Renamed reducer for invite
