+++ app/src/redux/reducer.js
@@ -5,7 +5,7 @@ import login from './login';
 import event from './eventReducers';
 import profile from './profileReducers';
 import eventsList from './eventsReducers';
-import invite from './invite';
+import invite from './inviteReducers';
 import newEvent from './newEventReducers';
 
 export default combineReducers({
